Backend/deploy/requirements.txt
# PyTorch and associated libraries
torch==2.2.2
tensorboard

# Hugging Face libraries
transformers==4.40.0
datasets==2.18.0
accelerate==0.29.3
evaluate==0.4.1
bitsandbytes==0.43.1
huggingface_hub==0.22.2
trl==0.8.6
peft==0.10.0

# Other libraries
Pillow==9.5.0
matplotlib==3.8.0
numpy==1.24.3
pandas==2.1.0
tqdm==4.66.1
pytesseract
langdetect
torchvision

# flask server and stuff
Flask==2.3.3
Flask-Cors==4.0.0
waitress==3.0.0
Flask-HTTPAuth